+++ assets/js/utils/time.tsx
@@ -213,6 +213,14 @@ export function endOfCurrentYear() {
   return new Date(now.getFullYear(), 11, 31);
 }
 
+//
+// Calculates the precise time difference between two dates in terms of complete months, weeks, and days.
+// The function handles string or Date inputs and ensures calculation from earlier to later date.
+//
+// Example:
+//   getDateDifference("2025-01-10", "2025-02-20")
+//   returns: {months: 1, weeks: 1, days: 3}
+//
 export function getDateDifference(date1: string | Date, date2: string | Date) {
   const parsedDate1 = typeof date1 === "string" ? parse(date1) : date1;
   const parsedDate2 = typeof date2 === "string" ? parse(date2) : date2;
